refactor: Agent Marketplace Button with access control
