+++ client/src/components/Nav/NewChat.tsx
@@ -1,11 +1,10 @@
-import React, { useCallback, useContext } from 'react';
-import { LayoutGrid } from 'lucide-react';
+import React, { useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useQueryClient } from '@tanstack/react-query';
-import { QueryKeys, Constants, PermissionTypes, Permissions } from 'librechat-data-provider';
+import { QueryKeys, Constants } from 'librechat-data-provider';
 import { TooltipAnchor, NewChatIcon, MobileSidebar, Sidebar, Button } from '@librechat/client';
 import type { TMessage } from 'librechat-data-provider';
-import { useLocalize, useNewConvo, useHasAccess, AuthContext } from '~/hooks';
+import { useLocalize, useNewConvo } from '~/hooks';
 import store from '~/store';
 
 export default function NewChat({
@@ -27,15 +26,6 @@ export default function NewChat({
   const navigate = useNavigate();
   const localize = useLocalize();
   const { conversation } = store.useCreateConversationAtom(index);
-  const authContext = useContext(AuthContext);
-  const hasAccessToAgents = useHasAccess({
-    permissionType: PermissionTypes.AGENTS,
-    permission: Permissions.USE,
-  });
-  const hasAccessToMarketplace = useHasAccess({
-    permissionType: PermissionTypes.MARKETPLACE,
-    permission: Permissions.USE,
-  });
 
   const clickHandler: React.MouseEventHandler<HTMLButtonElement> = useCallback(
     (e) => {
@@ -57,21 +47,6 @@ export default function NewChat({
     [queryClient, conversation, newConvo, navigate, toggleNav, isSmallScreen],
   );
 
-  const handleAgentMarketplace = useCallback(() => {
-    navigate('/agents');
-    if (isSmallScreen) {
-      toggleNav();
-    }
-  }, [navigate, isSmallScreen, toggleNav]);
-
-  // Check if auth is ready (avoid race conditions)
-  const authReady =
-    authContext?.isAuthenticated !== undefined &&
-    (authContext?.isAuthenticated === false || authContext?.user !== undefined);
-
-  // Show agent marketplace when marketplace permission is enabled, auth is ready, and user has access to agents
-  const showAgentMarketplace = authReady && hasAccessToAgents && hasAccessToMarketplace;
-
   return (
     <>
       <div className="flex items-center justify-between py-[2px] md:py-2">
@@ -92,24 +67,6 @@ export default function NewChat({
           }
         />
         <div className="flex">
-          {showAgentMarketplace && (
-            <div className="flex">
-              <TooltipAnchor
-                description={localize('com_agents_marketplace')}
-                render={
-                  <Button
-                    variant="outline"
-                    data-testid="nav-agents-marketplace-button"
-                    aria-label={localize('com_agents_marketplace')}
-                    className="rounded-full border-none bg-transparent p-2 hover:bg-surface-hover md:rounded-xl"
-                    onClick={handleAgentMarketplace}
-                  >
-                    <LayoutGrid className="icon-md md:h-6 md:w-6" />
-                  </Button>
-                }
-              />
-            </div>
-          )}
           {headerButtons}
 
           <TooltipAnchor
